create a new util file
